+++ src/@lib/equalities/shallowEquals.ts
@@ -1,3 +1,24 @@
 export function shallowEquals<T>(objA: T, objB: T): boolean {
-  return objA === objB;
+  if (objA === objB) return true;
+
+  if (
+    objA === null ||
+    objB === null ||
+    typeof objA !== "object" ||
+    typeof objB !== "object"
+  ) {
+    return false;
+  }
+
+  const keysA = Object.keys(objA) as (keyof T)[];
+  const keysB = Object.keys(objB) as (keyof T)[];
+
+  if (keysA.length !== keysB.length) return false;
+
+  for (const key of keysA) {
+    if (!Object.prototype.hasOwnProperty.call(objB, key)) return false;
+    if (objA[key] !== objB[key]) return false;
+  }
+
+  return true;
 }
