+++ src/@lib/hocs/memo.ts
@@ -1,10 +1,22 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import { shallowEquals } from "../equalities";
-import { ComponentType } from "react";
+import { ComponentType, createElement, ReactElement, useRef } from "react";
 
 export function memo<P extends object>(
   Component: ComponentType<P>,
   _equals = shallowEquals,
 ) {
-  return Component;
+  return function MemorizedComponent(props: P): ReactElement {
+    const prevPropsRef = useRef<null | P>(null);
+    const prevRenderRef = useRef<ReactElement | null>(null);
+
+    const isUpdated =
+      !prevPropsRef.current || !_equals(prevPropsRef.current, props);
+
+    if (isUpdated) {
+      prevPropsRef.current = props;
+      prevRenderRef.current = createElement(Component, props);
+    }
+
+    return prevRenderRef.current!;
+  };
 }
