+++ src/@lib/hooks/useCallback.ts
@@ -1,10 +1,10 @@
-/* eslint-disable @typescript-eslint/no-unused-vars,@typescript-eslint/no-unsafe-function-type */
-import { DependencyList } from "react";
+/* eslint-disable @typescript-eslint/no-unsafe-function-type */
+import { DependencyList, useMemo } from "react";
 
 export function useCallback<T extends Function>(
   factory: T,
   _deps: DependencyList,
 ) {
   // 직접 작성한 useMemo를 통해서 만들어보세요.
-  return factory as T;
+  return useMemo(() => factory, _deps);
 }
