+++ src/@lib/hooks/useMemo.ts
@@ -1,12 +1,21 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import { DependencyList } from "react";
 import { shallowEquals } from "../equalities";
+import { useRef } from "./useRef";
 
 export function useMemo<T>(
   factory: () => T,
   _deps: DependencyList,
   _equals = shallowEquals,
 ): T {
-  // 직접 작성한 useRef를 통해서 만들어보세요.
-  return factory();
+  const prevDepsRef = useRef<DependencyList | null>(null);
+  const prevValueRef = useRef<T | null>(null);
+
+  const isUpdate = !prevDepsRef.current || !_equals(prevDepsRef.current, _deps);
+
+  if (isUpdate) {
+    prevDepsRef.current = _deps;
+    prevValueRef.current = factory();
+  }
+
+  return prevValueRef.current!;
 }
