+++ src/providers/notification/provider.tsx
@@ -1,0 +1,32 @@
+import { useCallback, useMemo, useState } from "react";
+import { NotificationContext } from "../../contexts";
+import { Notification } from "../../types";
+
+export const NotificationProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+  const addNotification = useCallback(
+    (message: string, type: Notification["type"]) => {
+      setNotifications((prev) => [...prev, { id: Date.now(), message, type }]);
+    },
+    [],
+  );
+
+  const removeNotification = useCallback((id: number) => {
+    setNotifications((prev) => prev.filter((n) => n.id !== id));
+  }, []);
+
+  const noticiationValue = useMemo(
+    () => ({ notifications, addNotification, removeNotification }),
+    [notifications, addNotification, removeNotification],
+  );
+
+  return (
+    <NotificationContext.Provider value={noticiationValue}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
