+++ src/providers/theme/provider.tsx
@@ -1,0 +1,23 @@
+import { useMemo, useState } from "react";
+import { ThemeContext } from "../../contexts";
+import { ThemeContextType } from "../../types";
+
+export const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
+  const [theme, setTheme] = useState("light");
+
+  const toggleTheme = () => {
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+  };
+
+  const ThemeValue: ThemeContextType = useMemo(
+    () => ({
+      theme,
+      toggleTheme,
+    }),
+    [theme],
+  );
+
+  return (
+    <ThemeContext.Provider value={ThemeValue}>{children}</ThemeContext.Provider>
+  );
+};
